+++ asp3_dcre/sample/tSample2.c
@@ -512,7 +512,7 @@ eMainTask_main(void)
 			consume_time(1000LU);
 			hrtcnt2 = fetchHighResolutionTimer();
 			syslog(LOG_NOTICE, "hrtcnt1 = %tu, hrtcnt2 = %tu",
-											hrtcnt1, hrtcnt2);
+								(uint32_t) hrtcnt1, (uint32_t) hrtcnt2);
 			break;
 
 		case 'v':
